+++ test/feature/extend-fn-feature.js
@@ -15,6 +15,36 @@ Feature('extendFn behaviour extender', () => {
     });
   });
 
+  Scenario('a timer start event', () => {
+    const behaviour = {
+      $type: 'bpmn:StartEvent',
+      eventDefinitions: [{ type: 'bpmn:TimerEventDefinition', behaviour: { timeCycle: '0 0 * * *' } }],
+    };
+
+    When('the behaviour is extended', () => {
+      extendFn(behaviour);
+    });
+
+    Then('the time cycle is lifted onto the behaviour as scheduledStart', () => {
+      expect(behaviour.scheduledStart).to.equal('0 0 * * *');
+    });
+  });
+
+  Scenario('a timer start event with a duration', () => {
+    const behaviour = {
+      $type: 'bpmn:StartEvent',
+      eventDefinitions: [{ type: 'bpmn:TimerEventDefinition', behaviour: { timeDuration: 'PT1H' } }],
+    };
+
+    When('the behaviour is extended', () => {
+      extendFn(behaviour);
+    });
+
+    Then('no scheduledStart is added - only a cycle is a schedule', () => {
+      expect(behaviour.scheduledStart).to.be.undefined;
+    });
+  });
+
   Scenario('a behaviour without extension elements', () => {
     const behaviour = {};